requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-allauth>=0.54.0
django-simple-captcha>=0.5.17
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.3.0
psycopg2-binary>=2.9.7
Pillow>=10.0.0
django-cors-headers>=4.3.0
schedule>=1.2.0
croniter>=1.4.1